requirements.txt
Django==1.8.6
djangorestframework==3.3.1
django-registration==2.0.4